Update Optimism Tutorial
This PR contains changes to the Optimism tutorial to bring it in line with the current state of the Optimism protocol. **Please note: additional changes may be needed when Optimism Bedrock goes live. I am not sure how that timing lines up with the timing of this course, as far as I know there is no official release date for Bedrock.**